requirements.txt
django
pillow
django-elasticsearch-dsl
psycopg2